requirements.txt
fastapi
uvicorn
asyncpg
python-dotenv
rapidfuzz
